Add files via upload
These are few of the generated face images. The images are conditioned on the age by taking same input vector and different ages with top left corner representing the age 1 year and bottom right representing the age of 60 year. The age is increasing from top left to bottom right. Clearly, we can see the age difference of the images generated by the GAN network.